requirements.txt
fastapi==0.111.0
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.1.0
uvicorn[standard]==0.30.0

sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0

passlib[bcrypt]==1.7.4
python-jose==3.3.0

pydantic==2.7.1
pydantic-settings==2.7.1
alembic==1.13.1

celery==5.3.6
redis==5.0.4

# Email and templates
jinja2==3.1.2
email-validator==2.1.0

# Logging
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
